ai_service/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
scikit-learn
requests
tensorflow
Pillow
